backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic
email-validator
python-multipart
httpx
